requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
openai==0.28.1
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
whitenoise==6.6.0
yarl==1.9.2
